Summary of Fixes Completed The Spanner Emulator type compatibility issues have been resolved. Here's what was fixed: Model Changes (internal/models/) int → int64 changes: care_plan.go: Version, ExpectedVersion medical_record.go: Version, ExpectedVersion, UsageCount (in MedicalRecordTemplate) medication_order.go: Version, ExpectedVersion social_profile.go: ProfileVersion visit_schedule.go: EstimatedDurationMinutes, PriorityScore sql.NullTime → spanner.NullTime: social_profile.go: ValidFrom, ValidTo time.Time → civil.Date: visit_schedule.go: VisitDate for DATE type compatibility Repository Changes (internal/repository/) Delete functions: Fixed to use single-column primary keys (not composite): care_plan_repository.go medication_order_repository.go visit_schedule_repository.go clinical_observation_repository.go DATE comparisons: Fixed GetUpcomingSchedules to use civil.Date instead of time.Time Test Changes (tests/integration/) int64 assertions: Changed assert.Equal(t, 60, ...) to assert.Equal(t, int64(60), ...) .String → .StringVal: Fixed spanner.NullString field access StatusOK → StatusNoContent: Fixed delete test expectations URL encoding: Fixed time series test to URL-encode date parameters Remaining Test Failures The remaining failures are not related to Spanner types: ValidationErrors tests: The API doesn't validate certain required fields - this is a validation implementation issue, not a database issue Template cleanup failures: Server closes before cleanup - cosmetic cleanup issue

+++ backend/internal/models/care_plan.go
@@ -27,7 +27,7 @@ type CarePlan struct {
 	Activities json.RawMessage `json:"activities,omitempty"`
 
 	// Optimistic Locking
-	Version int `json:"version"`
+	Version int64 `json:"version"`
 
 	CreatedBy string    `json:"created_by"`
 	CreatedAt time.Time `json:"created_at"`
@@ -65,7 +65,7 @@ type CarePlanUpdateRequest struct {
 	Goals      json.RawMessage `json:"goals,omitempty"`
 	Activities json.RawMessage `json:"activities,omitempty"`
 
-	ExpectedVersion *int `json:"expected_version,omitempty"` // Optimistic locking
+	ExpectedVersion *int64 `json:"expected_version,omitempty"` // Optimistic locking
 }
 
 // CarePlanFilter represents filter options for listing care plans
